requirements.txt
asgiref==3.8.1
attrs==24.3.0
Brotli==1.1.0
click==8.1.8
dj-database-url==2.3.0
Django==5.1.4
django-cors-headers==4.6.0
django-markdownx==4.0.7
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
drf-spectacular==0.28.0
drf-yasg==1.21.8
gunicorn==23.0.0
h11==0.14.0
inflection==0.5.1
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
Markdown==3.7
packaging==24.2
pillow==11.0.0
psycopg2-binary==2.9.10
PyJWT==2.10.1
pytz==2024.2
PyYAML==6.0.2
referencing==0.35.1
rpds-py==0.22.3
sqlparse==0.5.3
typing_extensions==4.12.2
uritemplate==4.1.1
uvicorn==0.34.0
whitenoise==6.8.2
